Reverting apt-update.yml split
The intial purpose of spliting up the tasks to apt-update.yml was to
force apt-get update handler to be run after the "block of tasks" in
that file (as mentioned in the Ansible handler documentation).
Unfortunately this is not the way it works. After some research found
flush_handlers directive which should do the job.