Allow the user to choose a different ntp implementation
Sometimes the default ntp package is not what we need, so we remove
'ntp' from the base Debian package selection and allow the user to
choose a different ntp implementation by setting the variable
'base_ntp_package'.

To avoid that no ntp package is installed, we still default to 'ntp'.
This might change in the future.

Add the variable + variable description to the "Default Role
Variables" section + also add some notes on NTP.

Thanks @mika for the report.

Closes: jkirk/ansible-role-base#7