Deploy Debian release specific chrony configuration
Debian/buster does not work with the default configuration file of
Debian/bullseye so we need to deploy a Debian release specific
configuration file.

While at it we also add a handler to restart chrony, if the
configuration has changed and we also make sure that the ntp service is
started.