Allow multiple allow directives
If chrony is installed on a Samba Domain Controller (DC), the location
of the Samba ntp_signd socket should be specified via
base_ntp_ntpsigndsocket / ntpsigndsocket.

It seems like ntpsigndsocket should not contain a trailing slash in
chrony (although ntp.conf used to have one).
Also see: https://wiki.samba.org/index.php/Time_Synchronisation#With_chrony

Removed the trailing slash in the examples and the README.md

To allow access to the server for MS-SNTP requests, we need to set the
allow directive. But we only were able to set a single allow directive.
('allow 127/8' was set by default because we want Prometheus
node_exporter to access the local timeserver).

Although a simple 'allow' allows access by any node (which /should/ be
accomplished by a `base_ntp_allow: ""` (untested), the better approach
is it to be able to allow multiple allow directives.

We now allow to set a list of allow directives via `base_ntp_allow`.