Update Github Workflows together with Ansible Molecule configuration
Inspired by geerlingguy's implementation of the CI workflow, decided to
give it a try.

Ansible Molecule only runs on Debian/bookworm by default, while the
Github workflow manages the distro matrix where Ansible Molecule runs on
Debian/buster, /bullseye and /bookwork.

See:

* https://github.com/geerlingguy/ansible-role-ntp/blob/master/.github/workflows/ci.yml
* https://github.com/geerlingguy/ansible-role-ntp/blob/master/molecule/default/converge.yml

Starting systemd services with the Github Runner still fails with
something like this:

  fatal: [instance]: FAILED! => {"changed": false, "msg": "Unable to
  start service chrony: Job for chrony.service failed because the
  control process exited with error code.\nSee \"systemctl status
  chrony.service\" and \"journalctl -xe\" for details.\n"}

Decided to set base_ntp_enable to false to overcome that problem.

Also changed the namespace to jkirk and this is what we currently use.
Without that change running molecule locally is not possible.

Finally updated the linter workflow to use the latest slim image and
removed the redundant superlinter workflow.